Add files via upload
A comparison of Docker-based and non-Docker-based development is shown below.

With Docker, we can develop, package, and run applications quickly:

🔹 The developers can write code locally and then build a Docker image and push it to a dev environment. In this way, the development work is shared within the team. The tests are run automatically in the dev environment.

An image is a read-only template with instructions for creating a Docker container. We use a Dockerfile to define the steps to create the image and run it.

🔹 The above process can run incrementally when bugs are found or improvements are needed.

🔹 When dev testing is complete, the Docker image is pushed to the production environment (often on the cloud).

Compared with traditional development without Docker, Docker is quite lightweight and fast, because only the changed part of Dockerfile is rebuilt every time we make a change.