Add files via upload
I’ve got good news 

In Kubernetes v1.28 Non-Graceful node shutdown is now in GA.

Let me explain what it means.

↳ The Problem:

Picture this: Sometimes, the power decides to take a little break, and our K8s node goes “Goodnight!” all of a sudden. 

This can be because of Hardware failure or unresponsive OS.

The node damage isn’t a problem for stateless applications but it’s a problem for stateful ones.

↳ The Reason:

In a Graceful Node Shutdown, Kubelet can detect a node shutdown event, and can handle the failover well.

During an unexpected shutdown, kubelet does not have time to properly terminate the pods and schedule back Pods on other nodes.

↳ The Fix with Non-Graceful Node Shutdown:

In such scenarios, you can add out-of-service taint on the Node. Use the following fix.

→ kubectl taint nodes <node-name> node[.]kubernetes[.]io/out-of-service=nodeshutdown:NoExecute

Note: Remove the “[“ and “]” from the above command.

This will do forceful pod deletion if no matching tolerations are found. Persistent volume failover can also be managed accordingly.

If this is useful, please follow me on GitHub!!